Crate dev_bestia_html_templating

dev_bestia_html_templating

Templating library for html in fullstack Rust, server-side or client-side in wasm
repository; version: 0.1.3 date: 2021-10-14 authors: Luciano Bestia

Try it

Motivation

Internet browsers are available practically on all platforms. They are based today on the W3C standards html5 and css3. All modern browsers support also wasm/webassembly (instead or beside javascript). Rust is a fantastic language to be compiled into wasm. This combination can be used as a cross-platform GUI.
A program user interface is made of data and presentation. Sadly html has this all mixed in one “file”. To separate data from presentation we use a templating library.
First a graphical designer creates a html5 + css3 presentation with some sample texts. When he is satisfied, then the programmer puts some comments and attributes inside the html file. These markers are processed by the templating library and the data is injected into the html. So we start with a clear separation of data and presentation and end with a normal html file.

Create a “standard” html page

The html page has to be MicroXml compatible, basically XHtml. Copy for example web_server_folder/review_list.html to a new html file. Open this file with a browser to preview it. I use the VSCode extension vscode-open-wsl and right-click on the file and Open with default application. In WSL2 I use my project wsl_open_browser. Now edit the html file to your liking and reload the browser with F5 to see the changes. Use some sample text to make it look as close to what you want. These texts will be later programmatically replaced, but they are invaluable while designing a nice layout.

Add markers

Inside the html you want to replace the sample texts with the data from the server. Before the text add the (invisible) marker for example <!--wt_crate_name-->. You can replace also an attribute if you insert an attribute before it like this data-wt_variable_name="next_attribute_name".
Now run the automation task cargo auto build that will copy/embed this file into files_mod.rs.

cargo crev reviews and advisory

We live in times of danger with supply chain attacks.
It is recommended to always use cargo-crev
to verify the trustworthiness of each of your dependencies.
Please, spread this info.
You can also read reviews quickly on the web:
https://web.crev.dev/rust-reviews/crates/

open-source free and free as a beer

My open-source projects are free and free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful,
please buy me a beer donating on my paypal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

Modules

html_template_mod
html templating library for html It can work on the server and on the client (web browser in wasm). it reads through every token in the html and when finds a template variable example: or data-wt_name=“class” or it calculates it and replace the next token: it can be a text_node, attribute, a complex node,… On the server the placeholders start with “s” like “st_name” : st_, sb_,… On the web web browser (client) the placeholders start with “w” like “wt_name”: wt_, wb,… It is possible to use partially on the server and partially on the client. It is compatible also with svg where special namespaces are used.

macros_mod

url_utf8_mod
url encoding and decoding in html This module is strictly limited to utf8 urls.
Url is made of parts, fragments or segments mostly delimited by slash “/”.
They must be separately encoded/decoded, not as a whole string.
It is impossible to guarantee that the whole string is correctly encoded/decoded.
But is possible to minimize the misuse of the String type for Url.
With the normal String it is not possible to force the developer to encode/decode.
With special wrapper types around String is possible to help the coder to write properly and not forget about it.
TODO: analyze if is possible to use more &str and Cow instead of always allocating String.
But urls are usually small and this is not a priority.

utils_mod
small utility functions

Macros

s
short macro s! for &str.to_string or format!(). because that is so common. Equivalents: String::new(), x.to_string(), x.to_owned(),…

url_u
returns UrlUtf8EncodedString::new_x Constructor macro for UrlUtf8EncodedString The attribute [macro_export] “moves” the macro in the main module. Macros cannot be inside impl like fn. The module names must be added to the code to work properly. TODO: use macro repetition to avoid having 4 different fn.