Module dev_bestia_html_templating::html_template_mod

html templating library for html It can work on the server and on the client (web browser in wasm). it reads through every token in the html and when finds a template variable example: or data-wt_name=“class” or it calculates it and replace the next token: it can be a text_node, attribute, a complex node,… On the server the placeholders start with “s” like “st_name” : st_, sb_,… On the web web browser (client) the placeholders start with “w” like “wt_name”: wt_, wb,… It is possible to use partially on the server and partially on the client. It is compatible also with svg where special namespaces are used.

Structs

Attribute
An attribute on a DOM node, such as id="my-thing"

ElementNode

PrefixForTemplateVariables

SubTemplate

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

Node

ServerOrClient
The same html templating can be used on the server or on the client but if we want to mix it, we need to distinguish template variables with different prefixes example: “st_” for server or “wt_” for client (web browser)

Traits

HtmlServerTemplateRender
the trait has only methods that can be implemented (overridden) It is not possible to define a method as “override forbidden” because the data model is always different and is known only to the project. if not implemented, the default functions below will be used.

Functions

decode_5_xml_control_characters
private fn - decode 5 xml control characters : “ ’ & < > https://www.liquid-technologies.com/XML/EscapingData.aspx I will ignore all html entities, to keep things simple, because all others characters can be written as utf-8 characters. it is mandatory that text is valid utf-8. https://www.tutorialspoint.com/html5/html5_entities.htm TODO: find a faster method // The standard library replace() function makes allocation,

decode_html_script_node
in html the