Trait dev_bestia_html_templating::html_template_mod::HtmlServerTemplateRender

pub trait HtmlServerTemplateRender {
    fn data_model_name(&self) -> String;

    fn replace_with_string(
        &self, 
        placeholder: &str, 
        subtemplate_name: &str, 
        row_number: usize
    ) -> String;

    fn render_html(&self, html: &str) -> String
    where
        Self: Sized,
    { ... }

    fn replace_with_url(
        &self, 
        placeholder: &str, 
        _subtemplate_name: &str, 
        _pos_cursor: usize
    ) -> UrlUtf8EncodedString { ... }

    fn retain_next_node_or_attribute(&self, placeholder: &str) -> bool { ... }

    fn replace_with_nodes(&self, placeholder: &str) -> Vec<Node> { ... }

    fn render_sub_template(
        &self, 
        template_name: &str, 
        _sub_templates: &Vec<SubTemplate>, 
        _prefixes: &PrefixForTemplateVariables
    ) -> Vec<Node> { ... }
}

the trait has only methods that can be implemented (overridden) It is not possible to define a method as “override forbidden” because the data model is always different and is known only to the project. if not implemented, the default functions below will be used.

Required methods

fn data_model_name(&self) -> String

name of data model for debugging

fn replace_with_string(
&self,
placeholder: &str,
subtemplate_name: &str,
row_number: usize
) -> String

returns a String to replace the next text-node or attribute value use macro s!() for a normal string

Provided methods

fn render_html(&self, html: &str) -> String where
Self: Sized,

The code for templating starts here. renders the complete html file. Not a sub-template/fragment.

fn replace_with_url(
&self,
placeholder: &str,
_subtemplate_name: &str,
_pos_cursor: usize
) -> UrlUtf8EncodedString

same as replace_with_string, but return url exclusively for attributes value of href and src the url must be encoded in the beginning because it encodes segments of url prior to being composed together. use macro url_u!() to create an url, very like format! I try to avoid String here to force the developer to not forget to url_encode

fn retain_next_node_or_attribute(&self, placeholder: &str) -> bool

boolean : is the next node rendered or not: “wb_” or “sb_”

fn replace_with_nodes(&self, placeholder: &str) -> Vec<Node>

returns a vector of Nodes to replace the next Node

fn render_sub_template(
&self,
template_name: &str,
_sub_templates: &Vec<SubTemplate>,
_prefixes: &PrefixForTemplateVariables
) -> Vec<Node>

renders sub-template